in principle active sync is working. now need to change the request rate
